General Cleanup, reorganization and a few fixes

+++ src/main/java/com/drevery/scpdisturbance/block/custom/FaceSkinFloor.java
@@ -9,15 +9,13 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.vector.Vector3d;
 import net.minecraft.world.World;
 
-public class FaceSkinFloor extends Block
-{
-    public FaceSkinFloor(AbstractBlock.Properties properties)
-    {
+public class FaceSkinFloor extends Block {
+    public FaceSkinFloor(AbstractBlock.Properties properties) {
         super(properties);
     }
-    public void onEntityCollision(BlockState state, World worldIn, BlockPos pos, Entity entityIn)
-    {
-        entityIn.setMotionMultiplier(state, new Vector3d(0.25D, (double) 1F, 0.25D));
+
+    public void onEntityCollision(BlockState state, World worldIn, BlockPos pos, Entity entityIn) {
+        entityIn.setMotionMultiplier(state, new Vector3d(0.25D, 1D, 0.25D));
         entityIn.attackEntityFrom(DamageSource.GENERIC, 4F);
     }
 }
